+++ app/Footer.tsx
@@ -1,0 +1,7 @@
+const Footer: React.FC = () => (
+  <div className='mx-auto my-10 mb-5 text-xs'>
+    Copyright © {new Date().getFullYear()}
+  </div>
+);
+
+export default Footer;
